Trying to fix asphaleia support

+++ ColorBadges.h
@@ -10,13 +10,13 @@
 
 @interface ColorBadges : NSObject
 + (instancetype)sharedInstance;
-+ (BOOL)isDarkColor:(int)color;
++ (BOOL)isDarkColor:(NSInteger)color;
 + (BOOL)areBordersEnabled;
 + (BOOL)isEnabled;
 
 // Return RGB ints. i.e. 0xRRGGBB.
-- (int)colorForImage:(UIImage *)image;
-- (int)colorForIcon:(id)icon; // Must be an SBIcon *
+- (NSInteger)colorForImage:(UIImage *)image;
+- (NSInteger)colorForIcon:(id)icon; // Must be an SBIcon *
 
 @end
